move CAMPAIGN_DATES
