+++ api/views/diagnostic.py
@@ -29,7 +29,7 @@ from common.utils import file_import, send_mail
 from data.models import Canteen, Teledeclaration
 from data.models.creation_source import CreationSource
 from data.models.diagnostic import Diagnostic
-from macantine.utils import CAMPAIGN_DATES, is_in_correction
+from data.models.diagnostic_teledeclaration_dates import CAMPAIGN_DATES, is_in_correction
 
 logger = logging.getLogger(__name__)
